fix file path for words, move GAME as class level variable, remove gamePlayController
