+++ src/main/java/com/aws/corona/charades/service/GameSetUpService.java
@@ -20,7 +20,7 @@ import java.util.Scanner;
 @Service
 public class GameSetUpService {
 
-    private final static String WORDS_FILE_PATH = "C:/source/Corona-Charades/src/main/java/com/aws/corona/charades/assets/words.txt";
+    private final static String WORDS_FILE_PATH = "C:\\source\\Corona-Charades\\src\\main\\java\\com\\aws\\corona\\charades\\assets\\words.txt";
 
     private Random r = new Random();
