+++ src/main/webapp/WEB-INF/views/player-names-team-one.jsp
@@ -17,11 +17,6 @@
       </div>
 
       <div>
-          <h2> Number of players in player form: </h2>
-          <h2> ${numOfPlayersInPlayerForm} </h2>
-      </div>
-
-      <div>
         <h3>Game Active Words List: </h3>
         <c:forEach items="${activeWords}" var="word">
             <h6>${word}</h6>
